+++ test/Reader/JavaPropertiesTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-config for the canonical source repository
- * @copyright Copyright (c) 2005-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-config/blob/master/LICENSE.md New BSD License
  */
 
@@ -112,5 +112,15 @@ ASSET;
     {
         $this->expectException(Exception\InvalidArgumentException::class);
         new JavaProperties($delimiter);
+    }
+
+    public function testProvidesOptionToTrimWhitespaceFromKeysAndValues()
+    {
+        $reader = new JavaProperties(JavaProperties::DELIMITER_DEFAULT, JavaProperties::WHITESPACE_TRIM);
+        $arrayJavaProperties = $reader->fromFile($this->getTestAssetPath('key-value-whitespace'));
+
+        $this->assertNotEmpty($arrayJavaProperties);
+        $this->assertEquals($arrayJavaProperties['single.line'], 'test');
+        $this->assertEquals($arrayJavaProperties['multiple'], 'line test');
     }
 }