+++ src/WriterPluginManager.php
@@ -17,34 +17,39 @@ class WriterPluginManager extends AbstractPluginManager
     protected $instanceOf = Writer\AbstractWriter::class;
 
     protected $aliases = [
-        'ini'      => Writer\Ini::class,
-        'Ini'      => Writer\Ini::class,
-        'json'     => Writer\Json::class,
-        'Json'     => Writer\Json::class,
-        'php'      => Writer\PhpArray::class,
-        'phparray' => Writer\PhpArray::class,
-        'phpArray' => Writer\PhpArray::class,
-        'PhpArray' => Writer\PhpArray::class,
-        'yaml'     => Writer\Yaml::class,
-        'Yaml'     => Writer\Yaml::class,
-        'xml'      => Writer\Xml::class,
-        'Xml'      => Writer\Xml::class,
+        'ini'            => Writer\Ini::class,
+        'Ini'            => Writer\Ini::class,
+        'json'           => Writer\Json::class,
+        'Json'           => Writer\Json::class,
+        'php'            => Writer\PhpArray::class,
+        'phparray'       => Writer\PhpArray::class,
+        'phpArray'       => Writer\PhpArray::class,
+        'PhpArray'       => Writer\PhpArray::class,
+        'yaml'           => Writer\Yaml::class,
+        'Yaml'           => Writer\Yaml::class,
+        'xml'            => Writer\Xml::class,
+        'Xml'            => Writer\Xml::class,
+        'javaproperties' => Writer\JavaProperties::class,
+        'javaProperties' => Writer\JavaProperties::class,
+        'JavaProperties' => Writer\JavaProperties::class,
     ];
 
     protected $factories = [
-        Writer\Ini::class      => InvokableFactory::class,
-        Writer\Json::class     => InvokableFactory::class,
-        Writer\PhpArray::class => InvokableFactory::class,
-        Writer\Yaml::class     => InvokableFactory::class,
-        Writer\Xml::class      => InvokableFactory::class,
+        Writer\Ini::class            => InvokableFactory::class,
+        Writer\JavaProperties::class => InvokableFactory::class,
+        Writer\Json::class           => InvokableFactory::class,
+        Writer\PhpArray::class       => InvokableFactory::class,
+        Writer\Yaml::class           => InvokableFactory::class,
+        Writer\Xml::class            => InvokableFactory::class,
         // Legacy (v2) due to alias resolution; canonical form of resolved
         // alias is used to look up the factory, while the non-normalized
         // resolved alias is used as the requested name passed to the factory.
-        'zendconfigwriterini'      => InvokableFactory::class,
-        'zendconfigwriterjson'     => InvokableFactory::class,
-        'zendconfigwriterphparray' => InvokableFactory::class,
-        'zendconfigwriteryaml'     => InvokableFactory::class,
-        'zendconfigwriterxml'      => InvokableFactory::class,
+        'zendconfigwriterini'            => InvokableFactory::class,
+        'zendconfigwriterjavaproperties' => InvokableFactory::class,
+        'zendconfigwriterjson'           => InvokableFactory::class,
+        'zendconfigwriterphparray'       => InvokableFactory::class,
+        'zendconfigwriteryaml'           => InvokableFactory::class,
+        'zendconfigwriterxml'            => InvokableFactory::class,
     ];
 
     /**