+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#42](https://github.com/zendframework/zend-config/pull/42) adds support for PHP 7.1 and 7.2.
 
 ### Deprecated
 
@@ -14,7 +14,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- Nothing.
+- [#42](https://github.com/zendframework/zend-config/pull/42) removes support for HHVM.
 
 ### Fixed